Handle many sections
From the manpage:

If the number of entries in the section header table is
larger than or equal to SHN_LORESERVE (0xff00), e_shnum
holds the value zero and the real number of entries in the
section header table is held in the sh_size member of the
initial entry in section header table.  Otherwise, the
sh_size member of the initial entry in the section header
table holds the value zero.