Use more relaxed memory store to flip "is_alive" flag
By default, std::atomic<T>::operator=() stores a given value with
the sequentially-consistent memory ordering, which is a very strong
memory order. We only need the atomicity of store in this case,
so we can use memory_order_relaxed instead.

This change improves performance when we have a lot of section
fragments.

This change shortens the time for garbage collection for clang-13
from 170ms to 100ms.