Disable TBB_STRICT when building TBB (on by default)
This allows mold to build under clang++-14 which catches the write-only
`intptr_t drained` in `tbb/arena.cpp` as an unused parameter that
ultimately triggers `-Werror` to error out and abort the build.

It's 2021 and there are still projects that default to `-Werror` as the
default instead of using it only in the dev and CI environments... smh.

Closes #107.