Revert "[ELF] Keep SHF_MERGE and SHF_STRINGS bits"
This reverts commit 4d78510cb07cc3a76dda950dabb3a9a321b4b7f8.

If the SHF_MERGE bit is set, sh_entsize must have the size of the unit
for that section. In the previous commit, we changed the behavior of
mold so that it sets SHF_MERGE bits in the output file. However, we
didn't set sh_entsize. So, the resulting section is technically invalid.

We merge sections with different sh_entsize into a single output
section, so there's no easy way to "keep" the original value.

In this commit, we simply revert the previous change so that mold won't
set SHF_MERGE in the output file.

https://github.com/rui314/mold/issues/155