[ELF] Optimize relocation processing for non-alloc sections
In mold, a relocation refers either a symbol or a section piece.
A section piece is a segment of a mergeable section such as
.rodata.str.1 or .debug_str.

Previously, we preprocessed all relocations referring mergeable
sections to find their corresponding section pieces. We did this
because we want to find a relocation target as quickly as possible
for `gc_sections` and `scan_rels`.

However, we didn't actually have to do that for non-alloc sections,
as non-alloc sections are not subject to neither `gc_sections` nor
`scan_rels`. So, we could skip relocation preprocessing for non-alloc
sections. This commit implement that optimization.

It looks like this is an effective optimization for programs that have
large debug info because debug sections tend to contain a lot of
relocations referring a .debug_str which is a mergeable section.
Here is a few notable examples.

            Output size  Before  After
  clang-14  2.2 GiB      1.455s  1.396s  (4% faster)
  mongodb   4.8 GiB      2.341s  1.925s  (17% faster)

These nubmers were measured on a simulated 16-core 32-thread machine.