[ELF] Rename .dynbss -> .copyrel
Currently, the sections we create for copy relocations are named .dynbss
and .dynbss.rel.ro. Section names shouldn't be significant in ELF
executables, so we should be able to name them as we want.

However, in practice, valgrind treat a section with the name .dynbss
in a special way and crashes if an executable contains both .bss and
.dynbss due to the following assertion failure:

  $ valgrind ./prog
  ==1960252== Memcheck, a memory error detector
  ==1960252== Copyright (C) 2002-2017, and GNU GPL'd, by Julian Seward et al.
  ==1960252== Using Valgrind-3.15.0 and LibVEX; rerun with -h for copyright info
  ==1960252== Command: ./prog
  ==1960252==

  valgrind: m_debuginfo/readelf.c:2271 (vgModuleLocal_read_elf_debug_info): Assertion 'di->bss_svma + di->bss_size == svma' failed.
  Segmentation fault (core dumped)

This commit is a workaround for the valgrind issue. We simply rename
dynbss and .dynbss.rel.ro to .copyrel and .copyrel.rel.ro, respectively.

Fixes https://github.com/rui314/mold/issues/159