Make virtual destructor public
I believe if you make the destructor a protected function, you can't
destruct it through a pointer to the base type. All destruction has
to be called through a pointer to a derived type.

There's no difference at the moment, but I think it is more conventional
to make a virtual destructor public.