try to find locations references to undefined symbols in debuginfo
When reporting undefined references, first try to search debuginfo
for the actual source location of where the reference comes from,
i.e. the line/column and possibly header file. DWARFv5 filenames
are not supported yet. Since debuginfo needs to be relocated to
be usable (otherwise various references between different debug
sections do not work), so the reporting now needs to be done
later (to be improved).

Signed-off-by: Luboš Luňák <l.lunak@centrum.cz>