Check whether NEEDED dsos contain undefined symbol.
The reason why #559 was failing is because the previous implementation
was not checking whether libraries mentioned in the dynamic section
have definitions for our dso's symbol. Technically we have to look up
libraries almost exactly like a loader, but in this patch I only look
at the RUNPATH + library paths mentioned on the command line.

Signed-off-by: Robert Bartlensky <bartlensky.robert@gmail.com>